+++ Controls/_grid/interface/IPropStorage.ts
@@ -7,7 +7,7 @@
 export interface IPropStorage {
    /**
     * @cfg {String} Уникальный идентификатор, по которому в хранилище данных будет сохранена конфигурация контрола.
-    * @remark В {@link /doc/platform/developmentapl/interface-development/controls/list/ списочных контролах}, в том числе {@link Controls/grid:SortingSelector}, такое хранилище используется для сохранения выбранной {@link /doc/platform/developmentapl/interface-development/controls/list/sorting/ сортировки}.
+    * @remark В {@link /doc/platform/developmentapl/interface-development/controls/list/ списочных контролах}, в том числе {@link Controls/sorting:Selector}, такое хранилище используется для сохранения выбранной {@link /doc/platform/developmentapl/interface-development/controls/list/sorting/ сортировки}.
     */
    propStorageId?: string;
 }
