+++ Controls/_sorting/Render/Selector.wml
@@ -37,7 +37,7 @@
                 on:dropDownClose="_resetSelectedArrow()"
                 dataLoadCallback="{{_saveLinkToItems}}">
             <ws:itemTemplate>
-                <ws:partial template="wml!Controls/_grid/Render/sortingSelector/SortingSelectorItemTemplate" scope="{{itemTemplate}}"></ws:partial>
+                <ws:partial template="wml!Controls/_sorting/Render/SelectorItemTemplate" scope="{{itemTemplate}}"></ws:partial>
             </ws:itemTemplate>
                 <ws:contentTemplate>
                     <ws:partial template="Controls/dropdown:inputDefaultContentTemplate">
