+++ Controls/grid.less
@@ -1,3 +1,2 @@
 @import 'Controls/_grid/_Grid';
 @import 'Controls/_grid/_SortingButton';
-@import 'Controls/_grid/_SortingSelector';
