+++ Controls/sorting.less
@@ -1,0 +1,1 @@
+@import 'Controls/_sorting/_selector';
