+++ Controls/sorting.ts
@@ -1,0 +1,9 @@
+/**
+ * Библиотека контролов для работы с сортировкой записей
+ * @library
+ * @includes ISortingSelector Controls/sorting:ISortingSelector
+ * @includes ISortingParam Controls/sorting:ISortingParam
+ * @public
+ * @author Аверкиев П.А.
+ */
+export {default as Selector} from 'Controls/_sorting/Selector';
